🧹refactor: Move Dashboard panels into seperate components
