+++ web-ui/src/App.tsx
@@ -1,4 +1,3 @@
-import React, { useEffect } from 'react';
 import { Provider } from 'react-redux';
 import { store } from './store';
 import Dashboard from './components/Dashboard';
